Moved widgets to their own source files and added nice constructors
